test fix

+++ project/vgmstream.cpp
@@ -1104,7 +1104,7 @@ void VGMStream::encodeNoteItem(const Project& project, const StreamNoteItem* ite
             if (item->on()) {
                 const PCMChannelSettings* pcmcs = dynamic_cast<const PCMChannelSettings*>(item->channelSettings());
                 int volume = pcmcs->volume() * item->note().velocity() / 100;
-                int att = 4 * (100 - volume) / 100;
+                int att = 4 * (float)(100 - volume) / 100;
                 data.append(0xF0 | item->channel());
                 data.append((quint8)att);
